# Conflicts:
#	backends/gdx-backends-gwt/src/com/badlogic/gdx/backends/gwt/GwtApplication.java
#	extensions/gdx-controllers/gdx-controllers-gwt/.classpath

+++ extensions/gdx-bullet/jni/src/bullet/BulletSoftBody/btSoftRigidDynamicsWorld.cpp
@@ -125,7 +125,7 @@ void	btSoftRigidDynamicsWorld::solveSoftBodiesConstraints( btScalar timeStep )
 
 }
 
-void	btSoftRigidDynamicsWorld::addSoftBody(btSoftBody* body,short int collisionFilterGroup,short int collisionFilterMask)
+void	btSoftRigidDynamicsWorld::addSoftBody(btSoftBody* body, int collisionFilterGroup, int collisionFilterMask)
 {
 	m_softBodies.push_back(body);
